uint8_t for line

+++ src/display.cpp
@@ -50,7 +50,7 @@ void display_clear_buffer(){
  * Add a "string" to the display buffer.
  *
  */
-void display_set_buffer_line(int line, const char *str){
+void display_set_buffer_line(uint8_t line, const char *str){
     int len = strlen(str);
     display_buffer[line] = (char *)realloc(display_buffer[line], len + 1);
     memcpy(display_buffer[line], str, len);
@@ -62,7 +62,7 @@ void display_set_buffer_line(int line, const char *str){
  * Add a "string" to the display buffer, except centered this time.
  *
  */
-void display_set_buffer_line_centered(int line, const char *str){
+void display_set_buffer_line_centered(uint8_t line, const char *str){
     int len = strlen(str);
     int spaces = (OLED_MAXLEN - len) / 2;
     display_buffer[line] = (char *)realloc(display_buffer[line], spaces + len + 1);
@@ -77,7 +77,7 @@ void display_set_buffer_line_centered(int line, const char *str){
  * Add a "string" to the display buffer based on some arbitrary data's start and length (i.e. PDUInfo).
  *
  */
-void display_set_buffer_line(int line, uint8_t *data, uint16_t start, uint16_t len) {
+void display_set_buffer_line(uint8_t line, uint8_t *data, uint16_t start, uint16_t len) {
     display_buffer[line] = (char *)realloc(display_buffer[line], len + 1);
     memcpy(display_buffer[line], &data[start], len);
     display_buffer[line][len] = '\0';
@@ -89,7 +89,7 @@ void display_set_buffer_line(int line, uint8_t *data, uint16_t start, uint16_t l
  * but this time with a statically defined prefix.
  *
  */
-void display_set_buffer_line(int line, const char *prefix, uint8_t *data, uint16_t start, uint16_t len) {
+void display_set_buffer_line(uint8_t line, const char *prefix, uint8_t *data, uint16_t start, uint16_t len) {
     int prefix_len = strlen(prefix);
     display_buffer[line] = (char *)realloc(display_buffer[line], prefix_len + len + 1);
     memcpy(display_buffer[line], prefix, prefix_len);
@@ -103,7 +103,7 @@ void display_set_buffer_line(int line, const char *prefix, uint8_t *data, uint16
  * but this time with a statically defined prefix - integer input edition.
  *
  */
-void display_set_buffer_line_int(int line, const char *prefix, uint8_t *data, uint16_t start, uint16_t len) {
+void display_set_buffer_line_int(uint8_t line, const char *prefix, uint8_t *data, uint16_t start, uint16_t len) {
     int prefix_len = strlen(prefix);
 
     // this is so jank, and i have no idea if it works
